\name{robustsae-package}
\alias{robustsae-package}
\docType{package}
\title{Robust Bayesian Small Area Estimation }

\description{The package provides a function \code{robustsae} for full non-subjective Bayesian analysis for the general area level small area models. This considers small area modeling of both the population means and the population variances. This is possible due to the availability of additional data purported to estimate the error variances. Also, in order to induce some robustness of the procedure, t-prior for the random effects is used. When the data set includes true values for interest parameter, this function returns the comparison criteria.}

\details{
\tabular{ll}{
Package: \tab robustsae\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2016-12-05\cr
License: \tab GPL-3\cr
}
This package provides function for full Bayesian analysis of small area models. 
}
\author{
Malay Ghosh, Jiyoun Myung, Fernando Moura

Maintainer: Jiyoun Myung <jiyoun@ufl.edu>
}

\references{
Chip, S., and Green berg, E. (1995). Understanding the Metropolis-Hastings Algorithm.
\emph{The American Statistician}, \bold{49}, 327-335.

Rao, J. N. K. (2003) \emph{ Small Area Estimation}. John Wiley and Sons.

You, Y. and Chapman, B. (2006) Small Area Estimation Using Area Level Models and Estimated Sampling 
Variances. \emph{ Survey Methodology,} \bold{ 32:} 97-103.

Malay Ghosh, Jiyoun Myung, and Fernando Moura. (submitted) Robust Bayesian Small Area Estimation.
}

